total_second should return a float

+++ tim-commons/tim_commons/__init__.py
@@ -1,2 +1,2 @@
 def total_seconds(td):
-  return (td.microseconds + (td.seconds + td.days * 24 * 3600) * 10 ** 6) / 10 ** 6
+  return float(td.microseconds + (td.seconds + td.days * 24 * 3600) * 10 ** 6) / 10 ** 6
